refactor(installer): centralize dependency resolution in hub
